+++ changelog.md
@@ -24,6 +24,7 @@
 - Fixed (none) text to be black and standard font
 - Adset number and text alignment adjusted
 - Lock faction board after 1.5 sec in TTS to work around floating boards
+- Fixed crafted improvements snap points
 
 ## [1.12.1] - 2026-5-05 Forge-Tuning